requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
redis==5.0.1
python-socketio==5.11.1
prometheus-client==0.19.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
torch==2.2.2
python-engineio==4.9.0
aiohttp==3.9.3
websockets==12.0
